Add lint.sh to run golanci-lint in 2 stages
First stage is to run goimports and formatter, second is full suite.

This ensures that imports and formatting are fixed even in presence of
other issues. Otherwise golanci-lint refuses to fix anything.

This helpful when running aider with config like this - aider will use that
to autofix what it can after every update:

% cat .aider.conf.yml
lint-cmd:
  - "go: ./lint.sh"