Set the write bit for files written during template initialization
This used to work because the permission bits for builtin templates were
hardcoded to 0644 for files and 0755 for directories.

As of #1912 (and the PRs it depends on), builtin templates are no longer
pre-materialized to a temporary directory and read directly from the embedded
filesystem. It turns out that this builtin filesystem returns 0444 as the
permission bits for the files it contains. These bits are carried over to the
destination filesystem.

This change updates template materialization to always set the owner's write
bit. It doesn't really make sense to write read-only files and expect users to
work with these files in a VCS (note: Git only stores the executable bit).

The regression shipped as part of v0.235.0 and will be fixed as of v0.238.0.