Disable setup-go caching except in tests
Otherwise all those jobs compete for the same key and build cache ends up disabled on ubuntu.

Also disable golangci-lint cache, it should be fast enough as is.

Disable cache on the releaser job, even though it takes care to use a different key - it makes sense
to have fresh build for release.