Disable cleanup in testcli package
The main CLI entry point used to be a global variable, so global state had to
be cleaned up after every test run. This hasn't been the case for a while and
instead the CLI is initialized in a function call. State accumulated by a
single CLI "instance" can no longer leak into other instances, so we no longer
have to perform cleanup.