Remove lint.sh; re-add 'make fmt'
See Makefile for explanation on difference between 'make fmt' and 'make lint'.

I also removed lint.sh. Original motivation was to use it in aider, but it's not
a good fit there, because aider passes filenames and it does not work well with
golangci-lint which requires whole packages to work.

Follow up to #2062.