acc: fix inheritance of null values
Previously, one could not set LocalOnly = true in parent directory and then override
it with 'LocalOnly = false' in child directory. This is because, 'false' is considered
empty value by mergo.

In order to distinguish between 'explicitly set to false' and 'not set', I've changed
all simple variables in config to be pointers. Now, one can always override those, because
non-nil pointers are not null (with mergo.WithoutDereference option).